Network refactoring (#172)
* addclient and addnetwork refactor
* modnetwork refactor
* dhcp-functions refactor
* bash-completions and help-workflows
* a lot of mini improvements
* release notes

+++ usr/share/drlm/client/del/default/259_del_client_user.sh
@@ -1,8 +1,10 @@
 # delclient workflow
 
-userdel -r $CLI_NAME &> /dev/null
-if [ $? -eq 0 ];then 
-  Log "Client user $CLI_NAME deleted from OS"
-else 
-  LogPrint "WARNING: Client user $CLI_NAME  can not be deleted from OS!"
+if getent passwd $CLI_NAME &> /dev/null; then
+  userdel -r $CLI_NAME &> /dev/null
+  if [ $? -eq 0 ];then 
+    Log "Client user $CLI_NAME deleted from OS"
+  else 
+    LogPrint "WARNING: Client user $CLI_NAME can not be deleted from OS!"
+  fi
 fi
